Refactor opcode schedulers and implement one based on the ortools TSP solver to minimize the cycle cost to visit all changes in our estimated list.
This is fortunately a tractable (though slow) computation that does give
improvements on the previous heuristic at the level of ~6% better
throughput.

This opcode schedule prefers to group by page and vary over content, so
implement a fast heuristic that does that.  This scheduler is within 2%
of the TSP solution.